// src/oncobridge-web/src/app/api/models/cancer-staging-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CodedConceptResponse } from '../models/coded-concept-response';
import { PartialDateResponse } from '../models/partial-date-response';
import { StageCategoryResponse } from '../models/stage-category-response';
export interface CancerStagingResponse {
  categories: Array<StageCategoryResponse>;
  effective?: (null | PartialDateResponse);
  id: string;
  method?: (null | CodedConceptResponse);
  patientId: string;
  primaryCancerDiagnosisId: string;
  stageGroup?: (null | CodedConceptResponse);
}