// src/oncobridge-web/src/app/api/models/cancer-surgical-procedure-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CodedConceptResponse } from '../models/coded-concept-response';
import { TemporalOccurrenceResponse } from '../models/temporal-occurrence-response';
export interface CancerSurgicalProcedureResponse {
  bodySite?: (null | CodedConceptResponse);
  code: CodedConceptResponse;
  id: string;
  patientId: string;
  performed?: (null | TemporalOccurrenceResponse);
}