// src/oncobridge-web/src/app/api/models/patient-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CodedConceptResponse } from '../models/coded-concept-response';
import { PartialDateResponse } from '../models/partial-date-response';
export interface PatientResponse {
  birthDate?: (null | PartialDateResponse);
  id: string;
  sexAtBirthAsRecorded?: (null | CodedConceptResponse);
  sourceIdentifier?: (string | null);
}