// src/oncobridge-web/src/app/api/models/primary-cancer-diagnosis-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CodedConceptResponse } from '../models/coded-concept-response';
import { PartialDateResponse } from '../models/partial-date-response';
import { TemporalOccurrenceResponse } from '../models/temporal-occurrence-response';
export interface PrimaryCancerDiagnosisResponse {
  bodySite?: (null | CodedConceptResponse);
  code: CodedConceptResponse;
  id: string;
  onset?: (null | TemporalOccurrenceResponse);
  patientId: string;
  recordedDate?: (null | PartialDateResponse);
}